Added support for svn merge
JavaHL already had support for this method, but it did not exist in ISVNClientAdapter so I had
to add it.  I also added an implementation for the command line adapter.  The JavaHL method does 
not support the --dry-run option, so that feature of merge is not currently available.

While implementing this, noticed that I missed adding a message resource for the Switch operation,
so that is in this commit as well.