Created Revert dialog
When you do the revert option, you are now presented with a dialog that shows all of the resources
that will be reverted.  You can uncheck items that you do not want to revert, and only the items
that are checked will be reverted.

So this is both a confirmation dialog, as well as additional functionality in that you can more easily
selectively revert certain resources.