This fixes incoming deletes not being removed from the Sync view after an update.
Debugging calculateKind() in SVNStatusSyncInfo, I saw that when a file
was deleted the SVNStatusKind for the local resource was SVNStatusSyncInfo.DELETED (4),
which makes sense.  After doing an update, it became SVNStatusSyncInfo.UNVERSIONED (5).

So, if the local resource does not exist, and the kind is SVNStatusSyncInfo.UNVERSIONED
(a combination that really doesn't make sense), I return SyncInfo.IN_SYNC.  Fixes the
incoming delete problem, and doesn't seem to have the potential to cause some other problem.