Applied and tested patch from jeichhorn.
When having serveral repositories on the same server and the repository name
starts with the same substring like:

1) https://server/repos
2) https://server/repos1
3) https://server/repos2 

only the first repository location is stored in the .svnProviderState file after
restarting eclipse. The repositories are written correctly into the state file,
but when loading the file, the check if the location is part of an already
existing repository url (in method getRepository(..) in the class
org.tigris.subversion.subclipse.core.repo.SVNRepositories) will return the first
SVNRepository instance when the the second or third is loaded.
 
When stopping eclipse, subclipse will write out the state file only with the
first repository.

The fix is to change the check so that the instance is
retured when the location equals the url of an existing repository or the
location is a subfolder of an existing url (like https://server/repo/trunk).

Issue #: 234