Changes to how SVN conflicts are handled.
When an SVN conflict is generated we now create an Eclipse problem marker for
it, with the Edit Conflicts action contributed as the Quick Fix to the problem.

This also causes a Warning icon to be displayed in the bottom left decorator
position, as well as to ripple up the folder hierarchy (unless there is also
a need to display an error indicator -- which would be common.

Changed the conflicted_ov.gif icon again.  This time so that it looked like
merged_ov.gif which is what CVS displays in this scenario.  This avoids the
problem of seeing two warning decorators side by side, and since since the
warning decorator is still associated via the marker, it makes the icons
for the Edit Conflicts and Mark Resolved actions make sense.

One final note -- for existing users, the new "SVN Conflict" problem marker
appears that it will be turned off by default in their problem filters.  So
you have to click the filters button and enable it to see the SVN Conflict
item in the Problems view.