Enhancements to adding svn properties.
Add dialog presents known valid properties in a drop-down list and can do
simple validations of the values, such as enforcing true/false or a number.

The list of valid properties can be contributed by an extension point, and that
is how the "built-in" properties are contributed.  When a property is
contributed, it can be limited to just files or just folders, or Both.

As part of this change, removed the stuff in the existing schema related to
groups.  While the idea makes a lot of sense, I could not envision the UI
ever being more complex than a drop-down, so having the groups in there was
just overkill.