Added a new preference to enable serializing the status cache to/from disk.  The preference is on by default.
This has the potentially to vastly speedup the time from starting Eclipse to having
your decorators appear.  This is especially true on Windows where the svn status 
command is much slower than on *nix and OS X.

On my system it takes about 6 seconds to start Eclipse and see your workspace.  I
have one large workspace with about 24 projects.  Before this enhancement, it would
take another 50+ seconds to see decorators.  With this enhancement, it takes only
2 seconds.  That seems to be about the time it takes on all platforms and workspaces --
2 seconds.  So that gives an idea on the speedup.

If the workspace contents have changed outside of Eclipse, then doing a Refresh (which
you have to do anyway) still correctly refreshes the cache.