Fix NPE in Show in Resource History when the user has entered an invalid repository root URL.
The code now catches the exception that is thrown, and changes the URL that
is used to the repository URL.  It also update the repository root URL to be
equal to the repository URL so that this does not keep happening.

The initial error will still show something in the SVN Console.