Fix problem that shows in Eclipse log about stuff failing to initialize within 5000ms.  Most often, this was the SVNOutoutConsole.
The real problem is/was that the core plug-in took too long to finish
from the start() method.  I surmised that this is likely due to the time
it takes to load the three client adapters.

I modified SVNClientManager so that it does not load any adapters until
the default from the preferences is set by the UI plugin, and then it only
loads that one adapter.  This seems to improve startup time and gets rid
of the error.

If you open the preference page, then all adapters are loaded so you can
change adapters.  Also, if no preference has ever been set, then all adapters
are loaded as before.

Also, added some safe-guards so someone using just the core plugin will
still get correct behavior.  Ideally, they should call the 
setSvnClientInterface() method to control the adapter they get, but if
they do not, we will still load all of the adapters.