Several enahancements to Resource History view:
1)  Added support for the --stop-on-copy option of svn log.
This is set by a toggle in the view drop-down menu.

2)  Added support for the --limit n feature of svn log.  
Instead of always getting all revisions, we can now get n
revisions.  The default is 100 and can be set in the prefs.
A Get Next n and Get All button is added to the view toolbar
to get more revisions.