Added a workaround for a JavaHL bug when requesting status information via DAV.  We are always getting back a 0 for ReposLastCmtRevision.
I changed the code to return a null in this scenario and then changed
our Synch routine to check for the null and fall back to the old
behavior of running getInfo() to get the rest of the info.  The
effect of this is to negate the performance improvements in the status command
when using JavaHL via DAV.  However, the way the workaround is coded,
if this bug is "situational" at all then not everyone will experience this.

JavaSVN does not have the problem.  Nor does JavaHL via svn://.