Show out of date folders in the Synch view
Requires a workaround in the client adapter because the status API returns
null for the lastChangedRevision for folders that are out of date but not
otherwise changed.  The workaround calls svn info against the repository for
every folder that has a null value for lastChangedRevision.  If/when the
API's are fixed, the workaround will disappear on its own.